update(signup): transfer email input to an individual component
